refactor: new home structure
