+++ src/routes/index.svelte
@@ -2,97 +2,55 @@
 	import type { Load } from '@sveltejs/kit';
 
 	export const load: Load = async ({ page, fetch }) => {
-		try {
-			const { query } = page;
-			const res = await fetch(`/games.json?${query.toString()}`);
+		const { query } = page;
+		const res = await fetch(`/spellen.json?${query.toString()}`);
 
-			if (res.ok) {
-				const games = await res.json();
-
-				return {
-					props: {
-						games,
-						query,
-						offline: res.statusText === 'offline'
-					}
-				};
-			}
+		if (res.ok) {
+			const games = await res.json();
 
 			return {
 				props: {
-					status: res.status,
-					error: await res.json(),
+					games,
 					query
 				}
 			};
-		} catch (error) {
-			return {
-				status: 500,
-				error
-			};
 		}
+
+		const { message } = await res.json();
+
+		return {
+			error: new Error(message),
+			props: {
+				query
+			}
+		};
 	};
 </script>
 
 <script lang="ts">
-	import GameList from '$lib/GameList/GameList.svelte';
-	import GameListFilter from '$lib/GameList/GameListFilter.svelte';
-	import { onMount } from 'svelte';
+	import GameListHighlighted from '$lib/GameList/GameListHighlighted.svelte';
+	import GameListCarousel from '$lib/GameList/GameListCarousel.svelte';
 	import type { Game } from '$lib/games';
 
-	export let offline;
 	export let games: Game[];
-	export let query;
-
-	onMount(() => {
-		if (!navigator.onLine) offline = true;
-		patchGames();
-	});
-
-	function patchGames() {
-		Promise.all(
-			games.map((game: Game) => {
-				return caches
-					.open('gamesCache')
-					.then((cache) => {
-						return cache.match(`/games/${game.slug}.json`);
-					})
-					.then((response: Response | undefined) => {
-						if (response) game.offline = true;
-						else game.offline = false;
-						return game;
-					});
-			})
-		)
-			.then((patchGames: Game[]) => (games = patchGames))
-			.then(() => {
-				if (offline) games = games.filter((game) => game.offline);
-			});
-	}
+	const newestGames = games.slice(Math.max(games.length - 4, 0));
 </script>
 
 <header>
-	<h1>John Doe</h1>
-	<h2>Goedemorgen!</h2>
+	<h1>Jungle Gym</h1>
 </header>
 
-{#if !offline}
-	<GameListFilter {query} />
-{/if}
-
-<GameList {games} {offline} />
+<GameListHighlighted />
+<GameListCarousel gamesData={newestGames} carouselTitle="Nieuwe spellen" />
+<GameListCarousel gamesData={games} carouselTitle="Favoriete spellen" />
 
 <style>
+	h1 {
+		text-align: center;
+	}
+
 	header {
 		font-family: var(--font-heading);
-		display: flex;
-		flex-direction: column-reverse;
 		padding: 2rem 0 1rem;
-	}
-	header h2 {
-		margin: 0;
-		color: var(--color-grey);
-		font-weight: 400;
-		font-size: 1em;
 	}
 </style>
