+++ src/routes/spellen/[game].json.ts
